Fix DPC++ tests failures and Coverity failures
